chore: move stories to their related components
