+++ packages/ui/src/components/__chakra-stories__/Image.stories.tsx
